Tercer commit de backend EnjoyEsports

+++ backend/controllers/equipoController.js
@@ -1,5 +1,7 @@
 import Equipo from '../models/Equipo.js';
 import Universidad from '../models/Universidad.js';
+import Categoria from '../models/Categoria.js';
+import Usuario from '../models/Usuario.js';
 
 //esta funcion me traer todos los equipos de la base de datos
 const obtenerEquipos = async (req, res) => {
@@ -20,20 +22,60 @@ const obtenerEquipo = async (req, res) => {
 
 //esta funcion crea un equipo nuevo en la base de datos
 const nuevoEquipo = async (req, res) => {
-  const equipo = new Equipo(req.body);
-  const { universidad } = req.body; // sacamos el id de la uni del formulario
-  const existeUni = await Universidad.findById(universidad); // buscamos la uni ingresada en el backend
+  const { universidad, categoria, teamName } = req.body; // sacamos el id de la uni del formulario
+
+  const existeUni = await Universidad.findById(universidad);
+  const existeEquipo = await Equipo.findOne({ teamName });
+  let existeCategoria = await Categoria.findById(
+    categoria.match(/^[0-9a-fA-F]{24}$/)
+  ).select('-createdAt -updatedAt -__v -titulo -imagen -universidades');
+
+  if (!existeCategoria) {
+    const error = new Error('No se encontro la categoria');
+    return res.status(400).json({ msg: error.message });
+  }
   if (!existeUni) {
-    // hacemos una verificacion donde si no esta la uni en el backend retornamos un errror
     const error = new Error('No se encontro la universidad');
     return res.status(400).json({ msg: error.message });
   }
-  req.body.capitan = req.usuario._id; // seteamos un capitan con el usuario que hizo el request
+
+  if (existeEquipo) {
+    const error = new Error('Equipo ya registrado');
+    return res.status(400).json({ msg: error.message });
+  }
+
+  req.body.capitan = await req.usuario._id; // seteamos un capitan con el usuario que hizo el request
+  const { miembros } = await req.body;
 
   try {
+    for (const miembro of miembros) {
+      let existenMiembros = await Usuario.findById(miembro).select(
+        '-password -confirmado -token -createdAt -updatedAt -__v -cordinador -admin -email '
+      );
+      if (existenMiembros.datos.Equipo !== undefined) {
+        if (existenMiembros.datos.Equipo !== null) {
+          const error = new Error(
+            `el jugador ${existenMiembros.nombre} ya esta registrado en un equipo`
+          );
+          return res.status(401).json({ msg: error.message });
+        }
+      }
+    }
+
     const equipoAlmacenado = await Equipo.create(req.body);
     existeUni.teams.push(equipoAlmacenado._id);
+    existeCategoria.equipos.push(equipoAlmacenado._id);
+    for (const miembro of miembros) {
+      let existenMiembros = await Usuario.findById(miembro).select(
+        '-password -confirmado -token -createdAt -updatedAt -__v -cordinador -admin -email '
+      );
+      existenMiembros.datos.Equipo = equipoAlmacenado._id;
+      existenMiembros.datos.Universidad = existeUni._id;
+      await existenMiembros.save();
+    }
     await existeUni.save();
+    await existeCategoria.save();
+
     res.json(equipoAlmacenado);
   } catch (err) {
     console.log(err);
@@ -43,13 +85,42 @@ const nuevoEquipo = async (req, res) => {
 const actualizarEquipo = async (req, res) => {
   const { id } = req.params;
   const equipo = await Equipo.findById(id);
+  const { miembros } = await req.body;
+
   if (!equipo) {
     return res.status(404).json({ msg: 'Equipo no Encontrado' });
   }
+  if (equipo.capitan.toString() !== req.usuario._id.toString()) {
+    const error = new Error('Solo el capitan puede eliminar el equipo');
+    return res.status(401).json({ msg: error.message });
+  }
+  // seteamos lo ingresado en el formulario y si no hay modificaciones se queda con lo que esta en la base de datos
   equipo.teamName = req.body.teamName || equipo.teamName;
   equipo.titulo = req.body.titulo || equipo.titulo;
   equipo.miembros = req.body.miembros || equipo.miembros;
   equipo.universidad = req.body.universidad || equipo.universidad;
+
+  for (const miembro of miembros) {
+    let existenMiembros = await Usuario.findById(miembro).select(
+      '-password -confirmado -token -createdAt -updatedAt -__v -cordinador -admin -email'
+    );
+
+    if (existenMiembros.datos.Equipo.toString() !== equipo._id.toString()) {
+      console.log(
+        `el jugador ${existenMiembros.nombre} es de este mismo equipo`
+      );
+      if (existenMiembros.datos.Equipo !== undefined) {
+        if (existenMiembros.datos.Equipo !== null) {
+          const error = new Error(
+            `el jugador ${existenMiembros.nombre} ya esta registrado en un equipo`
+          );
+          return res.status(401).json({ msg: error.message });
+        }
+      }
+    }
+    existenMiembros.datos.Equipo = equipo._id;
+    await existenMiembros.save();
+  }
 
   try {
     const equipoAlmacenado = await equipo.save();
@@ -60,14 +131,75 @@ const actualizarEquipo = async (req, res) => {
 };
 
 const eliminarEquipo = async (req, res) => {
-  const { id } = req.params;
-  const equipo = await Equipo.findById(id);
+  const { id } = req.params; // scamos el id de la url
+  const equipo = await Equipo.findById(id); // buscamos el id en la base de datos
+
   if (!equipo) {
+    // verificamos si no existe el equipo, si no existe mandamos un error
     return res.status(404).json({ msg: 'Equipo no Encontrado' });
   }
+  const { categoria, miembros, universidad } = equipo;
+
+  const existeUni = await Universidad.findById(universidad); // buscamos la uni ingresada en el backend
+  let existeCategoria = await Categoria.findById(categoria).select(
+    '-createdAt -updatedAt -__v -titulo -imagen -universidades'
+  );
+
+  if (equipo.capitan.toString() !== req.usuario._id.toString()) {
+    const error = new Error('Solo el capitan puede eliminar el equipo');
+    return res.status(401).json({ msg: error.message });
+  }
+
   try {
-    await equipo.deleteOne();
+    const unis = existeUni.teams.filter((obj) => {
+      return obj.toString() !== equipo._id.toString();
+    });
+    existeUni.teams = unis;
+
+    const categorias = existeCategoria.equipos.filter((obj) => {
+      return obj.toString() !== equipo._id.toString();
+    });
+
+    existeCategoria.equipos = categorias;
+    for (const miembro of miembros) {
+      let existenMiembros = await Usuario.findById(miembro).select(
+        '-password -confirmado -token -createdAt -updatedAt -__v -cordinador -admin -email '
+      );
+      existenMiembros.datos.Equipo = null;
+      existenMiembros.save();
+    }
+    await existeUni.save();
+    await existeCategoria.save();
+    await equipo.deleteOne(); // eliminamos el objeto del equipo
     res.json({ msg: 'Equipo Eliminado' });
+  } catch (error) {
+    console.log(error);
+  }
+};
+
+const salirEquipo = async (req, res) => {
+  const { id } = req.params; // scamos el id de la url
+  const equipo = await Equipo.findById(id); // buscamos el id en la base de datos
+  if (!equipo) {
+    // verificamos si no existe el equipo, si no existe mandamos un error
+    return res.status(404).json({ msg: 'Equipo no Encontrado' });
+  }
+
+  const usuario = req.usuario._id;
+  let buscarMiembro = await Usuario.findById(usuario).select(
+    '-password -confirmado -token -createdAt -updatedAt -__v -cordinador -admin -email '
+  );
+  buscarMiembro.datos.Equipo = null;
+  const usuarios = buscarMiembro.equipos.filter((obj) => {
+    return obj.toString() !== equipo._id.toString();
+  });
+  equipo.miemrbos.pop(usuario);
+  // todo: cambiar la logia para uqe funcione porque con el .pop no funcionara
+
+  try {
+    await buscarMiembro.save();
+    await equipo.save();
+    res.json({ msg: 'Saliste del equipo correctamente' });
   } catch (error) {
     console.log(error);
   }
@@ -79,4 +211,5 @@ export {
   nuevoEquipo,
   actualizarEquipo,
   eliminarEquipo,
+  salirEquipo,
 };
